make offline mode do more: closes #57
it will now try to search the local registry and cache for the crate and
use the latest version. if its not found then an error is presented to
the user.

a warning is emitted if this method was used to notify the user that the
version may not be the most current